src/main/webapp/app/entities/pessoajuridica/delete/pessoajuridica-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { PessoajuridicaService } from '../service/pessoajuridica.service';

import { PessoajuridicaDeleteDialogComponent } from './pessoajuridica-delete-dialog.component';

describe('Pessoajuridica Management Delete Component', () => {
  let comp: PessoajuridicaDeleteDialogComponent;
  let fixture: ComponentFixture<PessoajuridicaDeleteDialogComponent>;
  let service: PessoajuridicaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [PessoajuridicaDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(PessoajuridicaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(PessoajuridicaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(PessoajuridicaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
